+++ digiplan/map/calculations.py
@@ -320,9 +320,9 @@ def heat_demand_per_municipality_2045(simulation_id: int) -> pd.DataFrame:
     return demand
 
 
-def detailed_overview(simulation_id: int) -> pd.DataFrame:  # noqa: ARG001
+def ghg_reduction(simulation_id: int) -> pd.Series:
     """
-    Calculate data for detailed overview chart from simulation ID.
+    Calculate data for GHG reduction chart from simulation ID.
 
     Parameters
     ----------
@@ -331,15 +331,21 @@ def detailed_overview(simulation_id: int) -> pd.DataFrame:  # noqa: ARG001
 
     Returns
     -------
-    pandas.DataFrame
-        holding data for detailed overview chart
+    pandas.Series
+        holding data for GHG reduction chart
     """
-    # TODO(Hendrik): Calculate real data
-    # https://github.com/rl-institut-private/digiplan/issues/164
-    return pd.DataFrame(
-        data={"production": [300, 200, 200, 150, 520, 0], "consumption": [0, 0, 0, 0, 0, 1300]},
-        index=["wind", "pv_roof", "pv_ground", "biomass", "fossil", "consumption"],
+    renewables = renewable_electricity_production(simulation_id).sum()
+
+    results = get_results(
+        simulation_id,
+        {
+            "electricity_production": electricity_production,
+        },
     )
+    electricity_import = results["electricity_production"].loc[["ABW-electricity-import"]]
+    electricity_import.index = electricity_import.index.get_level_values(0)
+    electricity_import["ABW-renewables"] = renewables
+    return electricity_import * 1e-3
 
 
 def electricity_from_from_biomass(simulation_id: int) -> pd.Series:
@@ -538,15 +544,8 @@ def electricity_overview(simulation_id: int) -> pd.Series:
         simulation_id,
         {
             "electricity_demand": electricity_demand,
-            "electricity_production": electricity_production,
         },
     )
-    renewables = results["electricity_production"][
-        results["electricity_production"].index.get_level_values(0).isin(config.SIMULATION_RENEWABLES)
-    ]
-    renewables.index = renewables.index.get_level_values(0)
-    renewables = pd.concat([renewables, pd.Series(electricity_from_from_biomass(simulation_id), index=["ABW-biomass"])])
-
     demand = results["electricity_demand"][
         results["electricity_demand"].index.get_level_values(1).isin(config.SIMULATION_DEMANDS)
     ]
@@ -556,6 +555,9 @@ def electricity_overview(simulation_id: int) -> pd.Series:
     demand["ABW-electricity-demand_hh"] += electricity_heat_production_result["electricity_heat_demand_hh"]
     demand["ABW-electricity-demand_cts"] += electricity_heat_production_result["electricity_heat_demand_cts"]
     demand["ABW-electricity-demand_ind"] += electricity_heat_production_result["electricity_heat_demand_ind"]
+
+    renewables = renewable_electricity_production(simulation_id)
+
     overview_data = pd.concat([renewables, demand])
     overview_data = overview_data.reindex(
         (
@@ -572,6 +574,22 @@ def electricity_overview(simulation_id: int) -> pd.Series:
     )
     overview_data = overview_data * 1e-3
     return overview_data
+
+
+def renewable_electricity_production(simulation_id: int) -> pd.Series:
+    """Return electricity production from renewables including biomass."""
+    results = get_results(
+        simulation_id,
+        {
+            "electricity_production": electricity_production,
+        },
+    )
+    renewables = results["electricity_production"][
+        results["electricity_production"].index.get_level_values(0).isin(config.SIMULATION_RENEWABLES)
+    ]
+    renewables.index = renewables.index.get_level_values(0)
+    renewables = pd.concat([renewables, pd.Series(electricity_from_from_biomass(simulation_id), index=["ABW-biomass"])])
+    return renewables
 
 
 def heat_overview(simulation_id: int) -> pd.Series: