+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 
 ## [Unreleased]
 ### Added
+- results for GHG reduction chart
 - adapt full load hours for renewables
 - demand results for 2045 scenario